Revert stackstac dtype to float64 to avoid fill_value incompatibility

+++ scripts/data/download_national_patches.py
@@ -265,8 +265,8 @@ def _download_source_month(
                 epsg=epsg,
                 resolution=10,
                 bounds=bounds_utm,
-                dtype=np.float32,
-                fill_value=0.0,
+                dtype=np.float64,
+                fill_value=0,
                 rescale=False,
             )
             if stack.shape[0] == 0 or stack.shape[1] == 0:
